panellayouter: use QTemporaryFile for applyLayout() (bsc#1213708, CVE-2023-32184)
This fixes a security issue, details of which will only be made
available once updates are in place in openSUSE.

Place the Python helper script into the file system and package it. Let
the script accept the path to a layout tarball as parameter. Write the
application resource into a QTemporaryFile to make it accessible to the
helper script.