[WIP] Completely rebuild the app in QML and overhaul the codebase
This commit will shred every speck of Jekyll out of the welcome
app, replacing it with a nice QML interface. While I'm not going
for a pixel-perfect recreation of the app, I'm trying to get a
UI that looks mostly the same as the Jekyll version.

As part of the rewrite, I'm overhauling the codebase to use best
practices for modern Qt and open-source projects in general. This
includes adding a .clang-format file to keep the code looking nice.

Closes #34